+++ src/main/java/insane96mcp/iguanatweaksreborn/module/experience/enchantments/integration/ToolBelt.java
@@ -1,5 +1,6 @@
 package insane96mcp.iguanatweaksreborn.module.experience.enchantments.integration;
 
+import dev.gigaherz.toolbelt.belt.ToolBeltItem;
 import dev.gigaherz.toolbelt.slot.BeltExtensionSlot;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.item.ItemStack;
@@ -11,8 +12,16 @@ public class ToolBelt {
     public static void putItems(List<ItemStack> items, LivingEntity livingEntity) {
         LazyOptional<BeltExtensionSlot> oBeltExtensionSlot = BeltExtensionSlot.get(livingEntity);
         oBeltExtensionSlot.ifPresent(beltExtensionSlot ->
-                beltExtensionSlot.getSlots().forEach(extensionSlot ->
-                        items.add(extensionSlot.getContents())
+                beltExtensionSlot.getSlots().forEach(extensionSlot -> {
+                            ItemStack belt = extensionSlot.getContents();
+                            if (belt.is(dev.gigaherz.toolbelt.ToolBelt.BELT.get())) {
+                                belt.getCapability(ToolBeltItem.ITEM_HANDLER).ifPresent(cap -> {
+                                    int slots = cap.getSlots();
+                                    for (int i = 0; i < slots; i++)
+                                        items.add(cap.getStackInSlot(i));
+                                });
+                            }
+                        }
                 )
         );
     }