+++ src/main/java/insane96mcp/iguanatweaksreborn/module/mobs/spawning/Spawning.java
@@ -39,7 +39,7 @@ import java.util.List;
 @LoadFeature(module = Modules.Ids.MOBS)
 public class Spawning extends Feature {
 
-    public static final SimpleBlockWithItem ECHO_LANTERN = SimpleBlockWithItem.register("echo_lantern", () -> new EchoLanternBlock(BlockBehaviour.Properties.copy(Blocks.SOUL_TORCH).lightLevel(state -> 7)));
+    public static final SimpleBlockWithItem ECHO_LANTERN = SimpleBlockWithItem.register("echo_lantern", () -> new EchoLanternBlock(BlockBehaviour.Properties.copy(Blocks.SOUL_LANTERN).lightLevel(state -> 7)));
     public static final RegistryObject<PoiType> ECHO_LANTERN_POI = ITRRegistries.POI_TYPES.register("echo_lantern", () -> new PoiType(ImmutableSet.copyOf(ECHO_LANTERN.block().get().getStateDefinition().getPossibleStates()), 1, 64));
 
     @Config