+++ changelog.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Upcoming
+## 3.10.6
 * Tentative balance
   * Swords now have -0.25 reach and lower knockback, but the sweeping attack now works even when running or jumping and deals the same knockback as the mob hit 
   * Shovels now reach farther (and fixed damage: was 2.5 base damage instead of 1 base damage +30%)